reorganization of mass attenuation material method
